Split support detail files into service card, service details, and generic interfaces
