+++ apps/super-app/src/services/ServiceSupportDetails.ts
@@ -1,0 +1,37 @@
+import { SupportDetail, FacebookLinksSection } from "./SupportDetail";
+import { LanguageOption } from "./Language";
+
+// We should consider moving TWC2-specific things to their own file.
+export interface Twc2SupportDetail extends SupportDetail {
+    blurbFacebookLinks: FacebookLinksSection;
+  }
+  
+  const TWC2_DETAIL: Record<LanguageOption, Twc2SupportDetail> = {
+    en: {
+      logo: '/images/twc2-logo.png',
+      name: 'Transient Workers Count Too (TWC2)',
+      website: 'https://twc2.org.sg/language-menu-for-qr-code/',
+      blurbIntro: `TWC2 helps all foreigners working in Singapore, including those on Work Permits and S-Passes. They help for free. Most common issues are salary non-payment, contract issues and injuries.`,
+      blurbFacebookLinks: {
+        title: `Join TWC2's Facebook Community Pages!`,
+        imageSrc: '/images/facebook-icon.svg',
+        links: [
+          {
+            href: 'https://www.facebook.com/twc2bangla',
+            text: 'Bengali বাংলা',
+          },
+          {
+            href: 'https://www.facebook.com/twc2tamil',
+            text: 'Tamil தமிழ்',
+          },
+        ],
+      },
+      ctaButtonText: '+65 6297 7564',
+      ctaLink: 'https://wa.me/6562977564',
+      ctaIcon: '/images/whatsapp-icon.svg',
+    },
+  };
+  
+  export function getTwc2Detail(language: LanguageOption): Twc2SupportDetail {
+    return TWC2_DETAIL[language];
+  }
